Major rewrite
Phew, this was a bit of work. Started modul... modularizing (??) the bot.  I made it modular!
Later on I think I'll make it load everything by default in the modules directory,
so you don't have to change anything in the code to add more modules.

Oh, and also made the bot use rich embeds for a lot of messages for really cool effects.
I still have more work to do, the latex stuff is still not a module, and I have more shit I have to add.

Oh also, most importantly - magic 8ball!!!!!